+++ .github/CONTRIBUTING.md
@@ -26,7 +26,7 @@ Install the dependencies locally before prior to development. You can use a virt
 
 ```console
 conda env create -f environment.yml
-conda activate tw-pywrap
+conda activate twkit
 ```
 
 You can then install the local repository for development using the `pip -e` command which will install it in place without copying it to your `PYTHONPATH`. Using `--no-deps` will ignore dependencies which have already been installed via Conda. This assumes the current working directory is a clone of this repository.
@@ -55,7 +55,7 @@ You can debug in VSCode for stepping through code and monitoring errors. To do s
       "name": "Python: Current File",
       "type": "python",
       "request": "launch",
-      "program": "${workspaceFolder}/tw-pywrap/cli.py",
+      "program": "${workspaceFolder}/twkit/cli.py",
       "args": ["-l", "debug", "example.yml"], // Change to your specific commands you wish to run
       "cwd": "${workspaceFolder}"
     }
@@ -73,7 +73,7 @@ You may need additional settings based on your set-up. A more complete example m
       "name": "Python: Current File",
       "type": "python",
       "request": "launch",
-      "program": "${workspaceFolder}/tw-pywrap/cli.py",
+      "program": "${workspaceFolder}/twkit/cli.py",
       "console": "integratedTerminal",
       "purpose": ["debugging"],
       "justMyCode": true,
