+++ examples/python/launch_hello_world.py
@@ -1,11 +1,11 @@
 import logging
 
-# Import the tw-pywrap package
-from tw_pywrap import tower
+# Import the twkit package
+from twkit import tower
 
 logging.basicConfig(level=logging.DEBUG)
 
-# Construct a new tw-pywrap Tower instance
+# Construct a new twkit Tower instance
 tw = tower.Tower()
 
 # Customise the entries below as required
@@ -13,7 +13,7 @@ workspace = "<YOUR_WORKSPACE>"  # Name of your Workspace
 compute_env = "<YOUR_COMPUTE_ENVIRONMENT>"  # Name of your Compute Environment
 
 # Specify a human-readable run name
-run_name = "hello-world-tw-pywrap"
+run_name = "hello-world-twkit"
 
 # Launch the 'hello-world' pipeline using the 'launch' method
 pipeline_run = tw.launch(
