--- tw_pywrap/__init__.py
@@ -1,6 +1,0 @@
-from warnings import warn
-
-warn(
-    "This package has been renamed to twkit. Please install twkit for future updates.",
-    DeprecationWarning,
-)
