+++ twkit/pipelines.py
@@ -2,7 +2,7 @@
 Subclass of Tower class for overriding pipelines subcommand methods.
 """
 from pathlib import Path
-from tw_py.tower import Tower
+from twkit.tower import Tower
 
 
 class Pipelines(Tower):
